+++ src/main/java/sch/cqre/api/repository/RentalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import sch.cqre.api.domain.RentalEntity;
 
 @Repository
-public interface RentalRepository extends JpaRepository<RentalEntity, Integer> {
-    RentalEntity findByBookIdAndUserId(Integer bookId, Integer userId);
+public interface RentalRepository extends JpaRepository<RentalEntity, Long> {
+    RentalEntity findByBookIdAndUserId(Long bookId, Long userId);
 }
